requirements.txt
python-telegram-bot
flask
pyTelegramBotAPI
gspread 
google-auth
apscheduler
python-dotenv
oauth2client
psutil
pytzopenai
